Just told Claude to fix shit n stuff

+++ kernel/src/include/krnl/libraries/std/string.h
@@ -37,7 +37,7 @@ uint64_t strlen(const char *str);
 void *memset(void *dest, int c, uint64_t n);
 void strncpy(char *dest, const char *src, uint64_t n);
 void *memcpy(void *dest, const void *src, uint64_t n);
-uint64_t memcmp(const void *dest, const void *src, uint64_t n);
+int memcmp(const void *dest, const void *src, uint64_t n);
 int strcmp(const char *s1, const char *s2);
 int strncmp(const char *s1, const char *s2, uint64_t n);
 int zerocheck(const void *dest, uint64_t n);
